Bug #1508516 ANDROID: cgroup: Implement Optionally killable freezer
[root cause  ] Implement Optionally killable freezer
[changes     ] Ggroup
[side effects] NA
[self test   ]:freezer
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:NA
[confirm dependent commit]:NA
[board]:common
[test case]:cgroup.freezer
[reviewers   ] Vincent.Wang
[change_type ] bringup--> sprd_chip_support
[tag_product ] common

cgroup: Implement Optionally killable freezer

Adds a configurable option to allow killing frozen tasks in a way
similar to cgroups v2. The status of frozen tasks is changed to
TASK_INTERRUPTIBLE and their PF_FROZEN flag is reset on pending fatal
signals.

This behavior is enabled via a run-time configurable option, so that
the pre-existing behavior is preserved by default.

Change-Id: Ic86a2c2652ab4b374cd636faf48bc55065858b3f
Bug: b/143308662
Test: manual
Signed-off-by: Marco Ballesio <balejs@google.com>
Signed-off-by: Li Guanglei <guanglei.li@unisoc.com>